test: llm fastapi code
