+++ experiments/llms/test/qwen_data/results_qwen.json
@@ -1,0 +1,242 @@
+[
+  {
+    "text": "철수한테 오천 원 보내줘",
+    "intent": "transfer",
+    "slots": {
+      "recipient": "엄마",
+      "amount": 10000
+    },
+    "response": "",
+    "_meta": {
+      "inference_time": 1.32
+    }
+  },
+  {
+    "text": "엄마한테 3만원만",
+    "intent": "transfer",
+    "slots": {
+      "recipient": "엄마",
+      "amount": 30000
+    },
+    "response": "",
+    "_meta": {
+      "inference_time": 1.34
+    }
+  },
+  {
+    "text": "만원만 보내줘",
+    "intent": "transfer",
+    "slots": {
+      "recipient": "엄마",
+      "amount": 1
+    },
+    "response": "",
+    "_meta": {
+      "inference_time": 1.1
+    }
+  },
+  {
+    "text": "엄마한테 보내줘",
+    "intent": "transfer",
+    "slots": {
+      "recipient": "엄마",
+      "amount": 30000
+    },
+    "response": "",
+    "_meta": {
+      "inference_time": 1.29
+    }
+  },
+  {
+    "text": "나 송금하고 싶어",
+    "intent": "transfer",
+    "slots": {
+      "recipient": "엄마",
+      "amount": 30000
+    },
+    "response": "",
+    "_meta": {
+      "inference_time": 1.28
+    }
+  },
+  {
+    "text": "가스비 내야 하는데 어떻게 해야 해?",
+    "intent": "other",
+    "slots": {
+      "recipient": "사람 이름",
+      "amount": null
+    },
+    "response": "",
+    "_meta": {
+      "inference_time": 0.87
+    }
+  },
+  {
+    "text": "내 계좌에 얼마 남았어?",
+    "intent": "inquiry",
+    "slots": {
+      "recipient": "나",
+      "amount": null
+    },
+    "response": "",
+    "_meta": {
+      "inference_time": 0.78
+    }
+  },
+  {
+    "text": "그 사람한테 또 보내줘",
+    "intent": "transfer",
+    "slots": {
+      "recipient": "제가 가진 돈",
+      "amount": 5000
+    },
+    "response": "",
+    "_meta": {
+      "inference_time": 1.46
+    }
+  },
+  {
+    "text": "아까랑 똑같이 보내줘",
+    "intent": "transfer",
+    "slots": {
+      "recipient": "엄마",
+      "amount": 30000
+    },
+    "response": "",
+    "_meta": {
+      "inference_time": 1.34
+    }
+  },
+  {
+    "text": "송금할래",
+    "intent": "transfer",
+    "slots": {
+      "recipient": "엄마",
+      "amount": 30000
+    },
+    "response": "",
+    "_meta": {
+      "inference_time": 1.38
+    }
+  },
+  {
+    "text": "이체 수수료 얼마야?",
+    "intent": "other",
+    "slots": {
+      "recipient": "저희",
+      "amount": null
+    },
+    "response": "",
+    "_meta": {
+      "inference_time": 1.0
+    }
+  },
+  {
+    "text": "오늘 아빠랑 점심 먹었어",
+    "intent": "inquiry",
+    "slots": {
+      "recipient": "아빠",
+      "amount": null
+    },
+    "response": "",
+    "_meta": {
+      "inference_time": 1.3
+    }
+  },
+  {
+    "text": "아빠한테 이만원 보내줘",
+    "intent": "transfer",
+    "slots": {
+      "recipient": "아빠",
+      "amount": 30000
+    },
+    "response": "",
+    "_meta": {
+      "inference_time": 1.37
+    }
+  },
+  {
+    "text": "여보한테 십만 원 이체해줘",
+    "intent": "transfer",
+    "slots": {
+      "recipient": "엄마",
+      "amount": 10000
+    },
+    "response": "",
+    "_meta": {
+      "inference_time": 1.21
+    }
+  },
+  {
+    "text": "보내지 마",
+    "intent": "other",
+    "slots": {
+      "recipient": null,
+      "amount": null
+    },
+    "response": "",
+    "_meta": {
+      "inference_time": 0.84
+    }
+  },
+  {
+    "text": "그냥 넘어가줘",
+    "intent": "other",
+    "slots": {
+      "recipient": null,
+      "amount": null
+    },
+    "response": "",
+    "_meta": {
+      "inference_time": 0.68
+    }
+  },
+  {
+    "text": "엄마 송금해줘",
+    "intent": "transfer",
+    "slots": {
+      "recipient": "엄마",
+      "amount": 30000
+    },
+    "response": "",
+    "_meta": {
+      "inference_time": 1.38
+    }
+  },
+  {
+    "text": "아, 삼만원 보내는 거였지",
+    "intent": "transfer",
+    "slots": {
+      "recipient": "엄마",
+      "amount": 30000
+    },
+    "response": "",
+    "_meta": {
+      "inference_time": 1.4
+    }
+  },
+  {
+    "text": "예금해줘",
+    "intent": "transfer",
+    "slots": {
+      "recipient": "부모님",
+      "amount": 50000
+    },
+    "response": "",
+    "_meta": {
+      "inference_time": 1.4
+    }
+  },
+  {
+    "text": "누구한테 보낼까요?",
+    "intent": "other",
+    "slots": {
+      "recipient": "알려주실 수 있는 누군가",
+      "amount": null
+    },
+    "response": "",
+    "_meta": {
+      "inference_time": 1.25
+    }
+  }
+]
